pyproject.toml
[project]
name = "audio-analytics"
version = "0.1.0"
description = "A microservice that transcribes audio files into text using Faster-Whisper ASR models."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "faster-whisper>=1.1.1",
    "kafka-python>=2.1.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
    "pytest>=7.0.0",
    "pytest-dotenv>=0.5.2",
    "coverage>=7.2.0",
    "httpx>=0.24.0", # for async FastAPI testing
    "pytest-cov>=6.1.1",
]

[tool.ruff]
line-length = 100  # or 88, depending on the preference
target-version = "py311"  # Python version
fix = true  # allow autofixing when running ruff

# Enable checks
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort import sorting
]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]  
show_missing = true
fail_under = 80
